feat(ProjectTier): implement new tier and restructure files
